Split MatchSubScreen into abstract component + timeshift implementation
